+++ tourcount/src/main/java/com/wmstein/tourcount/CountOptionsLActivity.java
@@ -185,7 +185,7 @@ public class CountOptionsLActivity extends AppCompatActivity implements SharedPr
         // automatically handle clicks on the Home/Up button, so long
         // as you specify a parent activity in AndroidManifest.xml.
         int id = item.getItemId();
-        if (id == R.id.home)
+        if (id == android.R.id.home)
         {
             Intent intent = NavUtils.getParentActivityIntent(this);
             assert intent != null;
@@ -276,7 +276,7 @@ public class CountOptionsLActivity extends AppCompatActivity implements SharedPr
                 {
                     url = new URL(urlString);
                     RetrieveAddr getXML = new RetrieveAddr(getApplicationContext());
-                    getXML.execute(url);
+                    getXML.doInBackground(url);
                 } catch (IOException e)
                 {
                     // do nothing