+++ docs/History.md
@@ -2,6 +2,13 @@
 
 #### TourCount Copyright (C) 2016-2023, Wilhelm Stein
 
+#### Version 3.3.5 (2023-05-08)
+- Gradle Plugin upgraded -> 8.0.0
+- Gradle upgraded -> 8.0
+- Deprecated startActivityForResult() replaced by ActivityResultLauncher
+- Deprecated execute(URL) replaced by onPostExecute(URL)
+- Some code improvements
+
 #### Version 3.3.4 (2023-03-26)
 - Gradle Plugin upgraded -> 7.4.2
 - Gradle upgraded -> 7.5